Reset optind to 1 before parsing options
Normally the system is already doing this for us. In the MPI simgrid
execution case, however, depending on the enabled optimizations for memory,
libc may not be privatized, to save the memory that would be needed for its
duplication, and to avoid the potential headaches brought by several libcs
running in the same process. The first process would then inrease optind,
and other processes would start with optind already increased.